fix: themes

+++ apps/api/src/modules/themes/themes.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, Query } from "@nestjs/common";
+import { BadRequestException, Body, Controller, Delete, Get, Param, Post, Put, Query } from "@nestjs/common";
 import { CreateThemeDto } from "./dto/create-theme.dto";
 import { ThemesService } from "./themes.service";
 import { UpdateThemeDto } from "./dto/update-theme.dto";
@@ -17,7 +17,10 @@ export class ThemesController {
 		return await this.themeService.createTheme({
 			name: dto.name,
 			year: +dto.year,
-			price: +dto.price
+			price: +dto.price,
+			organization: {
+				connect: { id: dto.organizationId }
+			}
 		})
 	}
 
@@ -30,24 +33,28 @@ export class ThemesController {
 			...(dto.name ? { name: dto.name } : {}),
 			...(dto.year ? { year: +dto.year } : {}),
 			...(dto.price ? { price: +dto.price } : {})
-
 		})
 	}
 
 	@Delete(":themeId")
 	async deleteTheme(
-		@Param("themeId") themeId: string
-
+		@Param("themeId") themeId: string,
+		@Query('organizationId') organizationId: string
 	) {
+		if (!organizationId) throw new BadRequestException();
+
 		return await this.themeService.deleteTheme(themeId);
 	}
 
 
 	@Get(":themeId")
 	async getTheme(
-		@Param("themeId") themeId: string
-
+		@Param("themeId") themeId: string,
+		@Query('organizationId') organizationId: string
 	) {
+
+		if (!organizationId) throw new BadRequestException();
+
 		return await this.themeService.getThemeByUniqueField('id', themeId);
 	}
 
@@ -55,14 +62,15 @@ export class ThemesController {
 	@Get()
 	async getAllTheme(
 		@Query() query: any
-
 	) {
+		if (!query.organizationId) throw new BadRequestException('pls add organizationId')
+
 		return await (
 			query.search ? this.themeService.searchInAllThemes(
 				query.search,
 				new PaginationDto(+(query.page || 1), +(query.limit || 50)),
-				{}
-			) : this.themeService.getAllTheme({},
+				{ organizationId: query.organizationId }
+			) : this.themeService.getAllTheme({ organizationId: query.organizationId },
 				new PaginationDto(+(query.page || 1), +(query.limit || 50)),
 			)
 		)
